Renamed file of model definition according to convention
