+++ frontend/src/app/modules/metric-finder/metric-finder.component.ts
@@ -1,6 +1,6 @@
 import {Component} from '@angular/core';
 import {Observable} from 'rxjs';
-import {TestResult} from './models/test-result';
+import {TestResult} from './models/test-result.model';
 import {MetricFinderService} from './services/metric-finder.service';
 
 @Component({
